+++ lib/domain/repository/image_repository_interface.dart
@@ -1,4 +1,4 @@
-import 'package:surf_flutter_summer_school_24/data/entity/image_entity.dart';
+import 'package:surf_flutter_summer_school_24/entity/image_entity.dart';
 
 abstract interface class IImageRepository {
   Future<List<ImageEntity>> getPhotos();
