+++ lib/entity/image_entity.dart
@@ -4,8 +4,9 @@ part 'image_entity.freezed.dart';
 
 @freezed
 class ImageEntity with _$ImageEntity {
-  const factory ImageEntity(
-      {required int id,
-      required String url,
-      DateTime? createdAt}) = _ImageEntity;
+  const factory ImageEntity({
+    required int id,
+    required String url,
+    DateTime? createdAt,
+  }) = _ImageEntity;
 }
