+++ lib/ui/pages/carusel_page/сarousel_page_model.dart
@@ -1,7 +1,7 @@
 import 'package:elementary/elementary.dart';
 import 'package:flutter/material.dart';
-import 'package:surf_flutter_summer_school_24/data/entity/image_entity.dart';
 import 'package:surf_flutter_summer_school_24/data/repositiry/mock_image_repository.dart';
+import 'package:surf_flutter_summer_school_24/entity/image_entity.dart';
 
 class CarouselPageModel extends ElementaryModel {
   final MockImageRepository imageRepository;
