+++ lib/ui/pages/grid_page/grid_page_widget.dart
@@ -2,7 +2,6 @@ import 'package:cached_network_image/cached_network_image.dart';
 import 'package:elementary/elementary.dart';
 import 'package:flutter/material.dart';
 import 'package:shimmer/shimmer.dart';
-import 'package:surf_flutter_summer_school_24/ui/pages/carusel_page/carusel_page_widget.dart';
 import 'package:surf_flutter_summer_school_24/ui/pages/grid_page/grid_page_wm.dart';
 import 'package:surf_flutter_summer_school_24/ui/ui_kit/widget/custom_app_bar.dart';
 
@@ -17,39 +16,33 @@ class GridPageScreen extends ElementaryWidget<GridPageWM> {
     return Scaffold(
       appBar: const CustomAppBar(),
       body: GridView.builder(
-          gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
-            crossAxisCount: 3,
-          ),
-          padding: const EdgeInsets.all(8),
-          itemCount: wm.images.value.length,
-          itemBuilder: (BuildContext context, int index) {
-            return Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 2.0, vertical: 3),
-              child: GestureDetector(
-                onTap: () {
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                      builder: (BuildContext context) => CarouselPageScreen(),
-                    ),
-                  );
-                },
-                child: CachedNetworkImage(
-                  imageUrl: wm.images.value[index].url,
-                  fit: BoxFit.cover,
-                  placeholder: (context, url) => Shimmer.fromColors(
-                    baseColor: Colors.black26,
-                    highlightColor: Colors.white24,
-                    child: Container(
-                      width: 116,
-                      height: 116,
-                      color: Colors.black26,
-                    ),
+        gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
+          crossAxisCount: 3,
+        ),
+        padding: const EdgeInsets.all(8),
+        itemCount: wm.images.value.length,
+        itemBuilder: (BuildContext context, int index) {
+          return Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 2.0, vertical: 3),
+            child: GestureDetector(
+              onTap: wm.openCarousel,
+              child: CachedNetworkImage(
+                imageUrl: wm.images.value[index].url,
+                fit: BoxFit.cover,
+                placeholder: (context, url) => Shimmer.fromColors(
+                  baseColor: Colors.black26,
+                  highlightColor: Colors.white24,
+                  child: Container(
+                    width: 116,
+                    height: 116,
+                    color: Colors.black26,
                   ),
                 ),
               ),
-            );
-          }),
+            ),
+          );
+        },
+      ),
     );
   }
 }
